Update meson usage for static windows builds and accomodate linux
If we use libplacebo as a subproject and set our meson options very
carefully then we can mostly let meson generate the link arguments for
us. The notable issue is that libplacebo will try to link shaderc and
spirv-cross-c-shared when we want shaderc_combined and spirv-cross-c.
This can be addressed by overriding the options. Note that this needs
glslang enabled as a dependency to get everything linked statically.

With this structure system libplacebo and dynamic linking should be
fully supported for linux users who want to do that.